+++ CHANGELOG.md
@@ -5,11 +5,47 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 3.1.0 - 2023-06-17
 
-* There will be a 3.x branch created for php 8.0, there is little differences between the functionality of the different
-  version of the package. The primary difference is the supported versions which have to do with the development of the
-  package and the dev dependencies.
+### Added
+
+* `functions.php` has been added and included in the autoload to make wrapper function available globally (#19)
+  * `dotGet()` has been added to wrap `Dot::Get()` static method (#19)
+  * `dotSet()` has been added to wrap `Dot::Set()` static method (#19)
+  * `dotHas()` has been added to wrap `Dot::Has()` static method (#19)
+  * `dotCount()` has been added to wrap `Dot::Count()` static method (#19)
+  * `dotDelete()` has been added to wrap `Dot::Delete()` static method (#19)
+  * `dotAppend()` has been added to wrap `Dot::Append()` static method (#19)
+  * `dotFlatten()` has been added to wrap `Dot::Flatten()` static method (#19)
+
+## 3.0.0 - 2023-06-16
+
+### Added
+
+* `Dot::get()`, `Dot::count()`, `Dot::delete()`, `Dot::append()` can now optionally throw
+  an `ryanwhowe\Dot\Exception\ArrayKeyNotSetException` if the requested searchKey is not set in the searchArray. (#17)
+
+### Changed
+
+* Require php `^8.0` (#17)
+* All instances of thrown `InvalidArgumentExcpetion` have been refactored to
+  be `ryanwhowe\Dot\Exception\InvalidDelimiterException` (#17)
+* Updated Exceptions thrown. All Exceptions are extended from the `ryanwhowe\Dot\Exception\DotException` class which is
+  extended from the base `\Expection` class. (#17)
+
+
+## 2.1.0 - 2023-06-17
+
+### Added
+
+* `functions.php` has been added and included in the autoload to make wrapper function available globally (#20)
+  * `dotGet()` has been added to wrap `Dot::Get()` static method (#20)
+  * `dotSet()` has been added to wrap `Dot::Set()` static method (#20)
+  * `dotHas()` has been added to wrap `Dot::Has()` static method (#20)
+  * `dotCount()` has been added to wrap `Dot::Count()` static method (#20)
+  * `dotDelete()` has been added to wrap `Dot::Delete()` static method (#20)
+  * `dotAppend()` has been added to wrap `Dot::Append()` static method (#20)
+  * `dotFlatten()` has been added to wrap `Dot::Flatten()` static method (#20)
 
 ## 2.0.0 - 2023-06-15
 
@@ -21,6 +57,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Require php `^7.1` (#4)
+
+## 1.2.0 - 2023-06-17
+
+### Added
+
+* `functions.php` has been added and included in the autoload to make wrapper function available globally (#21)
+    * `dotGet()` has been added to wrap `Dot::Get()` static method (#21)
+    * `dotSet()` has been added to wrap `Dot::Set()` static method (#21)
+    * `dotHas()` has been added to wrap `Dot::Has()` static method (#21)
+    * `dotCount()` has been added to wrap `Dot::Count()` static method (#21)
+    * `dotDelete()` has been added to wrap `Dot::Delete()` static method (#21)
+    * `dotAppend()` has been added to wrap `Dot::Append()` static method (#21)
+    * `dotFlatten()` has been added to wrap `Dot::Flatten()` static method (#21)
+
 
 ## 1.1.0 - 2023-06-13